Divided FTP server private classes to own classes
