+++ addons/binding/org.openhab.binding.networkcamera/src/main/java/org/openhab/binding/networkcamera/NetworkCameraBindingConstants.java
@@ -26,4 +26,5 @@ public class NetworkCameraBindingConstants {
     // List of all Channel ids
     public final static String IMAGE = "image";
     public final static String MOTION_TRIGGER = "motion-trigger";
+
 }
